+++ datastore/tests/unit/test_helpers.py
@@ -498,6 +498,41 @@ class Test_key_from_protobuf(unittest.TestCase):
         self.assertRaises(ValueError, self._call_fut, pb)
 
 
+class Test__get_read_options(unittest.TestCase):
+
+    def _call_fut(self, eventual, transaction_id):
+        from google.cloud.datastore.helpers import get_read_options
+
+        return get_read_options(eventual, transaction_id)
+
+    def test_eventual_w_transaction(self):
+        with self.assertRaises(ValueError):
+            self._call_fut(True, b'123')
+
+    def test_eventual_wo_transaction(self):
+        from google.cloud.proto.datastore.v1 import datastore_pb2
+
+        read_options = self._call_fut(True, None)
+        expected = datastore_pb2.ReadOptions(
+            read_consistency=datastore_pb2.ReadOptions.EVENTUAL)
+        self.assertEqual(read_options, expected)
+
+    def test_default_w_transaction(self):
+        from google.cloud.proto.datastore.v1 import datastore_pb2
+
+        txn_id = b'123abc-easy-as'
+        read_options = self._call_fut(False, txn_id)
+        expected = datastore_pb2.ReadOptions(transaction=txn_id)
+        self.assertEqual(read_options, expected)
+
+    def test_default_wo_transaction(self):
+        from google.cloud.proto.datastore.v1 import datastore_pb2
+
+        read_options = self._call_fut(False, None)
+        expected = datastore_pb2.ReadOptions()
+        self.assertEqual(read_options, expected)
+
+
 class Test__pb_attr_value(unittest.TestCase):
 
     def _call_fut(self, val):
